requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
gunicorn==21.2.0
azure-identity==1.14.0
azure-keyvault-secrets==4.7.0
azure-monitor-opentelemetry==1.0.0
applicationinsights==0.11.10
